+++ resources/web/streaming_viewer/index.html
@@ -36,7 +36,7 @@
     <div class="webots-view-container">
       <webots-view></webots-view>
     </div>
-    <script type="module" src="https://cyberbotics.com/wwi/R2023b/WebotsView.js"></script>
+    <script type="module" src="../wwi/WebotsView.js"></script>
     <script src="setup_viewer.js"></script>
   </body>
 </html>
