+++ resources/web/wwi/nodes/WbPropeller.js
@@ -1,0 +1,40 @@
+import WbGroup from './WbBaseNode.js';
+
+export default class WbPropeller extends WbGroup {
+  #device;
+  constructor(id) {
+    super(id);
+    this.currentHelix = -1; // to switch between fast and slow helix
+    this.#device = [];
+  }
+
+  get device() {
+    return this.#device;
+  }
+
+  set device(device) {
+    this.#device = device;
+  }
+
+  postFinalize() {
+    super.postFinalize();
+
+    if (typeof this.children[1] !== 'undefined')
+      this.currentHelix = this.children[1].id;
+    else if (typeof this.children[0] !== 'undefined')
+      this.currentHelix = this.children[0].id;
+    this.switchHelix(this.currentHelix, true);
+  }
+
+  switchHelix(id, force) {
+    if (id !== this.currentHelix || force) {
+      this.currentHelix = id;
+      this.children.forEach(child => {
+        if (child.id === this.currentHelix)
+          _wr_node_set_visible(child.wrenNode, true);
+        else
+          _wr_node_set_visible(child.wrenNode, false);
+      });
+    }
+  }
+}
